Basic time-totaling added. ATM totals by UID and Code (how much mins used by a code by UID)
Current caveat: the totaling is done in python(2.7?)
This mean it is victim to the 2038 yr problem.
This can probably be fixed by upping to python 3+
Another idea is that I can use sqlite to calculate totals.
But this would mean I need to store the notes data
in a db in memory. Meaning twice.
It would also mean I have to do stuff in sql!