Initial add of SQLite3 procs for tcal_ev.tcl. Most procs will use a return code for which tcal_ev will act on. Still to do a build db command.
For dprod_view & dprod_build, the TCL AND Python similars, I need
to change to schema(?) so that there is an rowid primary key.
While it didn't seem useful for viewing (maybe?), it would
seem immensely useful for editing.
(a linear complexity either way but less work if sqlite knows that
the row to be edited/delete is unique (in theory: I don't know
SQLITE's inner workings as well as anyone))