Changed osnv::s from a stored string to an initial index. This means I don't have to intercept every keystroke,  preserving any natural behavior.
This also changed more button from calling 'pressed' to
now 'wri_rep'. The numpad still uses 'pressed' so that it may
complete the rep macro -- this means numbers can be repeated
while using the number-row/phy-keyboard (but not onscreen-numpad)

Fixed symbols not printing due to not being escaped properly.
Minor issue: Enter is not bound (for the reg macro)