Got the code col to work. No idea why it failed before, checking the diff for answers. Changed from displaying end_time to the time-difference. (lesser chars and more use-able) ((Additional change from full year to 'YY; e.g. 1994 to 94 etc))
Idea to change from:
	getnstr to char[] and then set record's string
To:
	getnstr to record's internal char[] for the string

In reading on this, with c++11 this should be ok?

Also calling reserve on uid&code since their sizes are fixed & small.

number input yet to be fixed!. Still possible to take 0len input.
Will try to fix this along with adding a
range check (max & and logical like month -> days)