Made it work for C (sh7.c). You have to compile it with -std=c99 flag.
Interesting thing to note, when sh7 (and main) takes input
it only takes 1 number (as opposed to sh5 & sh6 that takes two)
The single number is $ROW$COL. Strangely this also corresponds to
the position in a 1d array. 01 -> row 0 col 1 or pos 1.
How this works with N_WIDTH!=10 I have no clue.