Made a test patterns file. Not used ATM. Made a 'FloodBoard' struct to move creation out of main. Still need work on tile_fill. FloodBoard.bsurf is equally distributed (except last color which may have 1 extra).
Going to change the fills to take a Floodboard instead of an
array, although with the annoyance from templates, I might change
from array to vector atleast temporarily. Maybe if I can find a way
to pass any container? The requirements are that it has [] operator
and size(). (size() wouldn't be need if I can store size in FloodBoard).

Alas.