Changed dummy field for RecordEditor to a 'save' button. Enter/Esc while in this field will respectively save or abort changes.
Changed 'new' on command bar to 'insert'.

Changed unsigned to size_t in cases when it is used to loop over arrays/
containers.

Added macro 'return_break' which just returns.
Replaced 'return' while in switch context to 'return_break'
Removed 'break' from above context.
This is just to remind that the entire function will return implying
a break. (but not needing it code-wise)