Initial add cfill_al.py.
Implemented basic cboard generation. (really easy in python with numpy.random.randint)
Implemented queue based bucket fill using 'west-east' (see floodalg.cpp)
Instead of a snitch func, this uses a 'changelog' var that is appended to if
it is not None. In later versions this will check if it is a list.

Differences:
	use of xy coord rather than singular 1d coord
	2d array as cboard as expected of above

scanline-fill (slfill) is TBD.

Really, if it works out well, there is nothing stopping me from
implementing cfill in Python rather than C++ as Python has a
curses lib with similarly named funcions.