va3 is an attempted implementation of a 3d matrix backed by a 1d array.
For a matrix of dims 2,3,4; the values are arranged as such:

[ 0              ][ 1              ] <-- z (2)

[ 0  ][ 1  ][ 2  ][ 0  ][ 1  ][ 2  ] <-- y (3)

[0123][0123][0123][0123][0123][0123] <-- x (4)

Unlike VecArr (the 2d one), this doesn't use a wrap_array but just calculates
the 1d index; this, however, means multi-indexing isn't implemented (yet)
To manipulate values with coordinates, use the 'at' function. This takes coords
in (z,y,x) order.

Due to the arrangement, looping through the base array (should?) will fill values
in cell-row-layer order. That is when increasing a 1d index: cells are filled when
the row is filled. when the row is filled, we proceed to the next row: which is
either of the same layer or the next layer. This loop is:
layer {
	row {
		col {}
	}
}

I just find this very neat and all.