For old stuff: changed types and header guards.
maze_int.cpp newly added.
My attempt towards 8 maze cells per int.
Made lots of realizations in making that.
For one: the layout is fliped horizontally and vertically.
E.g the cell I intended to be top left is bottom right.
While I could possiblely fix this, I realize that there
is little point. Who says that the first 4 bits has to be
the upper left of the block? It can equally be the bottom right.

((Also the realization that a 32bit int can hold 8 blocks /not/ 4.
I was also messing with shorts at the time..)).
This should not have any noticable effect. Only that the min maze
size went from 2x2 (WxH) to 2x4. The math should be the same as
I'm still dividing/modelo-ing by a multiple of 2.