Demo of ncurses print of a single cell. Really neato. Although I supposed since there are limited cases for a cell (2^4-1) I can just make some fixed cstring method? E.g MazePrint::printN(y,x). But that seems harsh.
Note to self: I used int instead of uint32_t here because
ncurses is C and since cell only uses 4 of the bits, changing
from unsigned to signed should do anything.

Wondering if it will be faster (or it is fast enough?) to print
the entire block at once? There would be less function calls but
I'll be more-or-less manually unrolling the code. We'll cross that
eventually.