bitstack.cpp added.
It is what the name implies, a stack for bits.
One may push n bits of an integer type on it, take top n bits,
and, (not implemented yet) pop off the top n bits.

ATM this is a simple demo of pushing and topping aswell as
how the bits are represented.

Internally (for now) I am using a vector of uint32.
There is little gain from using a uint32* as I did
in tvec.c (other than novelty)

TODO:
	change it to a class. e.g. BitStack.
	replace 'magic number' 32 with variable int size. (32/64..e.g)
	implement pop function