Same file
Allowed row deletion.
Allowed saving of modification time.
Added short comments on procs.
Reorganized menus.

Note:
Due to use of primary key on rowid, inserted rows will not be consecutive.

Possible fix (do I need it?):
Set it to autoincrement only and vaccumm on close.
Flaw with this is alteration of file regardless of changes.