Added hierarch_clo.tcl
This attempts to implement a hierarchy via closure table using
sqlite3 in TCL/TK (as opposed to naive ideas in hierarchy_naive.tcl)

I tried to move the code to bkmks.tcl but that would require an
overhaul of some sorts. So the commits to bkmks.tcl will break it
as the SQL does not match.

hierarch_clo developement is going.
When it is done, I will use include it in bkmks.tcl
rather than re-implement it.

The basic functions are made in hierarch although I have
not made a way to change subgroups (uplevel / downlevel).