bkmkg.tcl changes (and hierarch_clo.tcl)
Made/changed modify_row to include the group property which uses
a combobox to select a link's parent group or a checkbutton for a
lack of.

ATM changing groups via modify_row work visually (gui) but have not
looked at if it is changed at db level.
A update_data proc has been added to hierarch.

Made a proc for modifying a group (only gui elements ATM).