bkmkg.tcl
Made a save_grp proc that saves change to group
(like name and parent group)

It appears to work as intended BUT,
	due to the way groups ids are made "g$group.rowid"
	and root being {}, I have to branch when the parent
	group is root.

I'll try somethings later to change this.

Another issue is the use of checkbutton to mean root.
I really do not want to add 'root' to the groups combobox as
I think it is 1: a special group and 2: then you can't have
'root' as an actual group.
What's done work (at least at this point)