bkmkg.tcl Added a 'batch' group change function. While changing groups was possible through modify, one would have to select a new group for each link. The batch modify allows to set all selected links to the same group, regardless of it is already in that group.
In the process, generalized the 'group selection frame' to single
proc. I also think I'm going to rename these toplevels to the
same names since:
	1. only 1 can be open at any time
	2. many functions depend on the name and
		having the same one drops that dependency

Other than the bindings which I'm constantly forgeting to
put in, I plan to make some kind of 'sort/seperate' function
that just seperates the groups and data (recursively and visually).

Oh and that status bar.