bkmkg.tcl
Column sort (re)worked.
Firstly, it now a single one that switches to reorder_#0 if needed.
This is possible due to..
Second, not recursive. The old one is left in but, this one uses
'queue-like' tracking.

Still no (and not planned) multi sorting (ordered sort? by col 2,3)

So the differences between this and bkmks.tcl are:
	recursive or not.
	increasing/descreasing order or not.
	loaded as sorted or not.
	seperate index sort.

I say to each its own. If anything only the non-recursive part
is worth working for.