Baby steps; Deleted obsolete code (more on that below) Fixed broken constructor. Code is kept on bitbucket anyway + I have it written down.
A major change is the inclusion of a root node.
Before, the WordBank object kept any Node* array
and branched outwards foreach alphabet. Now it just
has one Node* that branches.
This change allows me to just push the root node on
to a queue/stack when traversing the tree. I don't
don't have "memory" I've lost by this change but
it makes coding easier.