Changed conflicting function argument and function name (prefix). Fixed (again) the no-argument run.
Made 'child_node' member function for Node that also traverses to new child.

Made two operator[] overload for Node* that
	1: takes a char ('a' to 'z' or 'A' to 'Z')
	2: takes a size_t (0 to 25)
and returns a Node* to that child.
(and nullptr if that doesn't exist)

Both child_node & the operator[] have safety precautions on invalid letters,
e.g non alpha-numeric. WordBank::new_word created as alternative to add_word
to take advantage. (add_word may be removed later on).

Some if tests were changed to use the [] operator as it removes repetition of
the l->p[cInd(s[i])]. Tests for ==nullptr and !=nullptr were change to !p and p
repectively. (since nullptr is false and not nullptr is true)

Some lines were reordered for readability where it did not change the function.