Made copy constructor, assignment operator, and augmented assign. Assignmented operator replaces the copy_ws. Augmented assign just acts as new_word with no length arg
In doing the above, I move copying from WordBank to Node's
scope with Node::copyr().

Attempted a prefix_subset function, although not working
ATM. What this is meant to do is given a prefix, return
a WordBank with words that have that prefix (as opposed
to a vector).

While the logic I put down seems sound, it doesn't work.
More later.