Added + operator for WordBank. Changed from 26 char to 36 char. This changes much, replacing cInd and cLow with tr. The NSIZE var is used for the index loop. node_walk has been disabled since there were unknown issues with it after the change. This was replaced with use of node_child.
cnt_word.cpp was changed to taste more functions due to
the prior breaking of prefix_subset. (it should work now.)

<< and += were swapped. That is, << will add words
while += will copy WordBank to another.
The reason being it matched the new + op and that
it would be good to add multiple words without using
parenthesises.

A NSeq struct was added to replace the old char i='a' to
'z' loops due to there 0-9 and a-z having a gap.
NSeq acts as a iterator of sorts.

add_child, has_child, and child_c (or whatever the names
were) are removed.