Changed child_node to childc and nchild to childi. Made childc call childi.
Made the empty string a default argument for some functions
like with_prefix() so maybe words() might be removed later
on.

Added the with_prefix count and remove_words.

Made 'CSeq' which is more or less like 'NSeq' wherein
incrementing and dereferencing reveals the desired values.
The difference is that CSeq works for substrings, that is
string,end_index. Most likely not going to use it.
Only function that uses it is prefix(CSeq) and
new_word_z.

Noticed possible error where in remove_word I loop over
string [lc to 1] rather than lc-1 to 1.

Although I think this was intended and I should have put
a comment stating that. (Note to self: put a comment there)