Initial add of parted.cs.
Simple implementation of 'ste.cpp' in C# .. mostly?
Instead of obscurely reading variables, I used a int
to tell which 'type' of part it is, so no polymorphs,
and no PIMPL.

Differences in parts are:
        Plain takes one to three args which make
                STR-DELIM-STR ..
        RandomPool takes one to four args which make
                RANC1..RANC1-RANC2..RANC2-..
        RandomN takes one to two args which make
                RAND(a,b)
        SubStr takes three args which make
                LIT[a][b..c]

Some new, some old.
Notice that one has to explcitly add Parts by the type
they desire unlike in srstrgen(_pm). This is because the
before-mentioned program consisted of three parts,
BInc (literal generator), PartedString (part compiler),
and main (argument parser). As this just makes ste, it
would be missing most functionality.