Tried to make auto_group && search_data use a shared proc to generate the 'where' part but, regression has occured where the regex is being matched not exactly.
eg. before 'Omen' matched nothing
unless it was exactly omen
now it matches if 'Omen' was _in_ the string

there are problems with '.*' too.

so.

solutions in mind:
1. might be a string/list problem.
due to strings and lists being interchange able
in tcl, i might be inputing a list when i want
a string (unlikely?)
2. it worked before somehow. maybe i'm not
allowed to build a stmt outside the proc?
3. syntax errors?
4. sqlite3 things i'm not aware of.
5. use tcl's regex instead of sqlite3's
this might be slower due to be naive?
6. replace regex with the 'like' operator.
^^ not likely, I'm not going back.
7. maybe it is working correctly and i'm
expecting the unexpected.