Moved Jso code to jso.cpp/h. Deleted some functions for JSON and added dereference op so that operations can be called on the Obj. Made Jso::jso_out which is an iterative print (most of the work is done in adder).
The flaw in the initial version was that I was trying to end
the loop with i=e (setting the itr to the end point)
But the loop condition check i!=e is checked after ++i meaning
i goes out of bounds and i==e is never true. another way I could
have did it was make a loop bool?