Bug: depends on input, not closing a brace may lead to seggy-fault despite efforts with exceptions.
I'm thinking it might be the use of chread's get since advance
increments dumbly leading to indices out of bounds but
attempts to change this by returning a '\0' leads to infinite looping
(by next_symplex?) despite putting a case for '\0'.

Also moved printing to jso.cpp and removed old printing.
Even though jso_out worked, it was a mess to read.