Temporary fix to get output working with hierarch_clo.tcl / bkmkg.tcl. Basically instead of output a copy of the tables, it mimics the schema used in hierarch_clo; not that I want this to be permanant since the newer one is improved.
Update on the broken mtime: for whatever reason, seg fault occurs
_whenever_ I attempt to retrieve key_value lastModified. I may have
been attempting to deref a nullptr, but from the JSON printout,
lastModified is a valid property with a value. At the moment, id
is in its place.

Hack to bypass jso/jso_f.cpp's lack of null type: replace null with
"null" in the json file.. Yes, I know its a not a fix. Esp. since
it defeats the purpose of chunkreading if one has to iterate 1 pass
from the get-go. (not to mention this doesn't fix lack of bool)

ATM I have to work out the specifics in mind, but atleast am planing
an alternative function to get strings/maps from a key_value pair.
Use of the internal x union is ugly.