Made an IReaderFactory: which does factory stuff. (and RAII) Cleaned up makefile.
On the fence whether to make a JsoFactory that does RAII because
what if I wanted to delete a Jso? I couldn't since the factory would
delete a ptr to nothing at the end. If I can't delete, then I can't
use temporary objects very easily; I would need to know the index in
the vector. Maybe a hashset of ptrs? Then I would be able to remove
deleted ptrs easily enough, but then would that defeat the purpose?