Added the gtk_gui so there's actually something to see..
Change from treeview to nodeview reason being that although
there is a hierarchy, its really just important only for the
groups (which is reflected in the sql code).

If I would include a group filter bar, having the entire group
collapse would make it redundent.

So in doing this change I need to (or rather should):
	make some kind of way to sort by hierarchy. maybe some way
		of keeping depth.

	make a grouping? to keep all subgroups of X adjacent.

	show/store the names of the groups but keep the group
		column as the id. A possible solution would be storing
		the group as a 'path', like 'g1/g2/g3' and that would
		certainly help with sorting,grouping, and naming (less so
		naming). the downside of it being fairly long.
			what do I need to do? Add a prop to linknode called
			'group' which is ref obj containing the XElement
			and a string 'path' as members?