Added long form arguments to SrStrGen (the old version) and a help if no arguments were given.
I am not fixing/improving that project. It has long since been reworked
as SpecStrGen and really only remains to be remembered as my 'first real
java project'. The recent fixes were done so the code can be somewhat
readable. The original code was made as a literal translation from the C++
version. While most of it is funtioning, it has a bad feel to it.

To come: making a argument parser for SpecStrGen. Might do it similar to
how I remade it here except as its own class and also flags for the parts
too.