Initial add of full_dist.py
This is for planning my vacation days next year. Instead of trying to
minimize working days per month overall, this time I'll try to distribute
this days so that time between day offs are maximized.

A problem I'm thinking of is tie-breaking for that 5 day weeks, which I
think I should maybe create a variable that stores distance to closest
holiday/vacation (e.g. not just a weekend)

Don't know what will come of this.