Change in thought process (good thing I haven't really done anything yet!) Instead of view displaying anything via ncurses, view will just carry the infomation of the view. (the structure backing the window). In this way, app will hold many views and display 1 of them at anytime using a window it controls.
Changes in the commit being removal of display properties from View. User will
query the state property of a View and call getItem dependent on that. getItem
will take a function that takes a const char* and call appropriately.

Special care is needed with various JTypes.

This project is going exceedingly slow mostly due to work on another repo, but
there are some ideas on paper.