Made small indicator for array items (view.cpp).
Began work on interface, app.cpp/h

Note: why was unique_ptr removed in use from view.cpp/h?
View no longer possesses anything (at the moment), before it was
thought to control the window, but app controls that now.
Maybe in the future, instead of a plain Jso*, it may hold a
unique_ptr<Jso*>, so that it would be able to still to getItems, but:
JSON controls that lifetime as well