Fixed: more items gotten than line on screen and Fixed: view::nextPage returning false
these two bugs were due to View::setItemsPerPage being called with LINES
BEFORE initscr, in which case LINES was not initilized, allowing the value
to be very high.

It doesn't seem to make sense that App and View to hold seperate itemsPerPage values
so it might be better for the former to hold it and just pass it when drawing.