Made url filters, again without command line option.
In testing, noticed that certain combinations of options gave no output,
such as: using a groupFilter and then sorting- this was because the step prior to sorting returned a generator instead of a list. Using the sort already in the source only modifies an iterable (the output is the same as the input); so it was getting sorted but lead to the generator being exhausted before output.
This was fixed by enforcing that the data must be a solid list before sorting. (or alternately, have sort return a new list)