Bug fix: url filters now work.
This bug was due to how the lambdas were created: when the url lambda were made, they referenced 'kv' in the function body. The problem is that kv changed as more filters were being introduced, meaning that all filters used the latest 'kv' and not the one present when the lambda was created (i.e. the expected one).
The fix was moving 'kv' to another scope, in this case: another function call actually creates the lambda.