After (finally) reading about canvas transformations, I put it to use.. which ironically would have been useful in doing scatterplot.*. Instead of using strange voodoo math I could have used translate and scale to do more or less the same thing. (scatterplot.* is more-or-less a copy of drawxy.py then)
That said, scale would only be used to flip the y plane and translate only to move the x,y origin, thereby replacing the equivilent transformX/Y functions.
Previous scaling/viewpoint technique will still prove useful.

Will play around with this; not sure if this will be 'back-ported' into scatterplot.* but it is an interesting revelation.