Minor change, storing scaled y coordinate.
Just a thought, the use of translate moves to the origin, but the origin is not guaranteeded to appear on the graph.
E.g. if xview was [-1001, -500].
A modified idea would be to translate to bottom left corner instead- thus allowing subtracting yMin as offset instead of adding yMax.
This is in addition to using a flipped y plane. This would reduce complexity a bit.

In regards to the axis and rotation, it should be possible to:
 save the context state,
 translate and rotate,
 draw axis and tick numbers,
 and then restore context state.