Test using global scale & translation. Tested scaling with prevous calculation, downside being that points will be appear non-square: requiring that the size being rescaled to account. In addition, lines are also scaled, but line width is only a single dimension making the above fix not usable. Lastly, font size scaling doesn't seem to have a fix.
Instead of using a 0,0 as the origin (which may not always in the graph), tried using the left most point (x sorted).

The previous success in drawing numbers was not repeated.

At the very least translating to left most point is a good start (after it has been scaled by my calculations)
The use of scale might just be limited to (1,-1) to flip the y plane.