Previous bug (clicks not registering) explained:   it occurs to rows horizonally parallel to the side buttons. Possible solution?: keep buttons above canvas
Made new function: drawState; which obviously draws the state- of the field.
Although it can replace any draws, it should be used when changing the canvas
element style in anyway. e.g. dimensions, colors.., without making a fresh
board.
Would be useful if I were to make the board resizable.