The add to window button now asks for sub code which changes what happens: If the subcode has been used before, modify the message/note with new one; otherwise: create a new subdiv with message on page.
NOTE: the bookmarklet is not synced with these changes, to reduce number
of modifications to the file until it is in more complete state.
To change the function to bookmarklet, wrap beginNotate function body in
an anonymous function javascript call and replace call to 'exinfo' with
some code that extracts info from page.